abstracting transformers
